Fix namespace collisions and move all utils into the same namespace
