+++ griptape/core/__init__.py
@@ -1,0 +1,6 @@
+from .base_tool import BaseTool
+
+
+__all__ = [
+    "BaseTool"
+]
