+++ griptape/core/executors/docker_executor.py
@@ -4,7 +4,7 @@ from attr import define, field, Factory
 import docker
 from docker.errors import NotFound
 import griptape
-from griptape.core.tools import BaseTool
+from griptape.core import BaseTool
 from griptape.core.executors import BaseExecutor
 import stringcase
 import tempfile
