--- griptape/core/tools/__init__.py
@@ -1,10 +1,0 @@
-from .base_tool import BaseTool
-from .decorators import (
-    action
-)
-
-
-__all__ = [
-    "BaseTool",
-    "action"
-]
