+++ griptape/utils/manifest_validator.py
