+++ tests/unit/memory/utils/test_command_runner.py
@@ -1,4 +1,4 @@
-from griptape.core.utils import CommandRunner
+from griptape.utils import CommandRunner
 
 
 class TestCommandRunner:
